+++ backend/src/test/java/com/sophia/store/service/BasicServiceTest.java
@@ -55,7 +55,7 @@ class BasicServiceTest {
             Set<Integer> idSet = new HashSet<>();
             for (int j = 0; j < NumericUtils.randomInteger(5, 10); j++) {
                 FormulaVo formulaVo = new FormulaVo();
-                formulaVo.setCount(NumericUtils.randomInteger(1, 5));
+                formulaVo.setCount(NumericUtils.randomFloat(1, 5));
                 Integer materialId = NumericUtils.randomInteger(1, 100);
                 int size = idSet.size();
                 idSet.add(materialId);
@@ -80,7 +80,7 @@ class BasicServiceTest {
         Set<FormulaVo> formulaVos = new HashSet<>();
         for (int i = 0; i < 10; i++) {
             FormulaVo formulaVo = new FormulaVo();
-            formulaVo.setCount(5);
+            formulaVo.setCount(5f);
             formulaVo.setId(i + 1);
             formulaVo.setType(Constant.MATERIAL);
             formulaVos.add(formulaVo);
